Fix OOM when comparing large, very different, Strings
Casting calculation of "edit distance memory" in tryReduceStringSizes to
use longs.

The result of multiplying the length of two Strings can be larger than
the maximum value a Integer can hold before overflowing. In this case,
comparing large strings with big differences could cause an overflow.
When it did overflow, the comparison against MAX_EDIT_DISTANCE_MEMORY
would give an incorrect result, possibly causing large Strings to be
passed to EditDistance, causing an OutOfMemoryError.

This now closely matches the similar check in
renderAsFailedStringComparison.

+++ spock-core/src/main/java/org/spockframework/runtime/ExpressionInfoValueRenderer.java
@@ -130,7 +130,7 @@ public class ExpressionInfoValueRenderer {
     end1++;
     end2++;
 
-    if ((end1-commonStart) * (end2-commonStart) > MAX_EDIT_DISTANCE_MEMORY) {
+    if (((long) end1-commonStart) * (end2-commonStart) > MAX_EDIT_DISTANCE_MEMORY) {
       return "false\nStrings too large to calculate edit distance.";
     } else {
       // Check if we can add some context
